pyproject.toml
[tool.poetry]
name = "spring-mass-damper-ML"
version = "0.1.0"
description = "Production-style physics surrogate (spring-mass-damper) with FastAPI, Docker, MLflow, Prometheus"
authors = ["Your Name <you@example.com>"]
packages = [{ include = "src" }]


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.114.0"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
pydantic = "^2.7.0"
numpy = "^1.26.4"
scipy = "^1.13.1"
torch = "^2.2.0"
mlflow = "^2.14.1"
prometheus-fastapi-instrumentator = "^6.1.0"
pyyaml = "^6.0.1"
httpx = "^0.27.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.7"
ruff = "^0.6.8"


[tool.ruff]
line-length = 100
select = ["E", "F", "I"]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
